==> PredictingTumorSpecific_TCGA/2_tumorexpressed_straightforward.ipynb <==
# IDENTIFICATION OF NOVEL CLASSES OF NEOANTIGENS IN CANCER | Tumor expressed


import os, glob, re, gtfparse
import pandas as pd
import numpy as np
from Bio import SeqIO
from functools import reduce


GENERAL="/users/genomics/marta" # same as previous step
projects=["BLCA"] # more project can be added here
bash_projects = " ".join(projects)

GENOMEDIR="/genomics/users/marta/genomes"
GENOMEFASTA=GENOMEDIR+"/GRCh38/GRCh38.primary_assembly.genome.fa"

transcript_gene=pd.read_csv(os.path.join(GENOMEDIR,"transcript_gene_v41.txt"), skiprows=1, names=['gene_id','gene_name', 'gene_type'])

## Quantification with featureCounts

%%bash -s "$bash_projects" "$GENERAL"

#############################################
######comment or uncomment the necessary#####

paired_single="paired-end"
# paired_single="single-end"

# strand=firststrand
# strand=secondstrand
strand=unstranded
#############################################

for proj in $1; do
    sbatch scripts/2_tumorexpressed/featureCounts_decision_straightforward.sh $proj $GENERAL $paired_single $strand
done

### modify headers
for proj in projects:
    df=pd.read_csv(os.path.join(GENERAL,proj,"analysis/07_quantification/straightforward/CountsTable_genename.txt"), skiprows=1, sep="\t")
    df['Geneid']=df['Geneid'].str.split('.').str[0]
    filter_col = [col for col in df if col.startswith('/')]
    for col in filter_col:
        new_col=col.split("BAM_files/")[1]
        new_col=new_col.split("Aligned")[0]
        df.rename(columns={col:new_col}, inplace=True)
    df.rename(columns={'Geneid':'gene_name'}, inplace=True)
    df.to_csv(os.path.join(GENERAL,proj,"analysis/07_quantification/straightforward/CountsTable_genename.header.txt"),sep="\t", index=False)
    

Convert counts to TPM

from rna_seq_normalization import Normalization as Norm
from sklearn.preprocessing import StandardScaler

# get TPMs
for proj in projects:
    tableofcounts = pd.read_csv(os.path.join(GENERAL,proj,"analysis/07_quantification/straightforward/CountsTable_genename.header.txt"),sep="\t")
    length = tableofcounts['Length']
    genes = tableofcounts['gene_name']

    info = tableofcounts[["gene_name","Chr","Start","End","Strand","Length"]]
    # we are only interested in the columns with counts
    counts = tableofcounts
    counts.drop(["gene_name","Chr","Start","End","Strand","Length"],axis=1, inplace=True)
    # calculate TPMs
    tpm_df = Norm.tpm(counts, length)
    columnames = tpm_df.columns.tolist()

    tpms = pd.concat([genes,tpm_df], axis=1)
    tpms.to_csv(os.path.join(GENERAL,proj,"analysis/07_quantification/straightforward/TPMs_genenames.csv"), index=False)

    # add all info apart from gene_names\n",
    tpms_w_info = pd.concat([info,tpm_df], axis=1)
    tpms_w_info.to_csv(os.path.join(GENERAL,proj,"analysis/07_quantification/straightforward/TPMs_genenames_whole_information.csv"), index=False)
    print(len(tpms_w_info.columns))



## Tumor-expressed

for proj in projects:
    fc = pd.read_csv(os.path.join(GENERAL,proj,"analysis/07_quantification/straightforward/TPMs_genenames.csv"))
    tumor_transcripts = list()
    patients = pd.read_csv(os.path.join(GENERAL,proj,"results/paired_patients.csv"))

    for index,patient in patients.iterrows():
        patient_fc = fc[["gene_name",patient.iloc[2]]]

        tumor_patient_fc = patient_fc[patient_fc[patient.iloc[2]] >= 1 ]
        tumor_transcripts.extend(tumor_patient_fc.gene_name.values.tolist())

    tumor = fc[fc['gene_name'].isin(tumor_transcripts)]

    lncRNA_pseudo = ['lncRNA','processed_pseudogene']

    tumor_merged = tumor.merge(transcript_gene, on=['gene_name'], how="inner")

    lncRNA = tumor_merged[tumor_merged['gene_type'].isin(lncRNA_pseudo)]
    print("lncRNA: ",len(lncRNA))

    cds = tumor_merged[tumor_merged['gene_type'] == "protein_coding"]
    print("PROTEIN CODING: ",len(cds))

    tumor1FPKM = pd.concat([lncRNA, cds], ignore_index = True)
    tumor1FPKM.to_csv(os.path.join(GENERAL,proj,"analysis/07_quantification/straightforward/tumor_1TPM_table_of_counts.csv"),index=False)
    tumor1FPKM[['gene_id','gene_name','gene_type']].to_csv(os.path.join(GENERAL,proj,"analysis/07_quantification/straightforward/tumor_1TPM.csv"), index=False)

    tumor1FPKM

for proj in projects:
    print(proj)
    fc = pd.read_csv(os.path.join(GENERAL,proj,"analysis/07_quantification/straightforward/tumor_1TPM_table_of_counts.csv"))

    patients=pd.read_csv(os.path.join(GENERAL,proj,"results/paired_patients.csv"))

    for index,patient in patients.iterrows():
        patient_fc = fc[["gene_name",patient.iloc[2]]]
        patient_fc = patient_fc[patient_fc[patient.iloc[2]] >= 1]
        filename=str(patient.iloc[0])+"_tumor1TPM.csv"
        try:
            os.makedirs(os.path.join(GENERAL,proj,"analysis/07_quantification/straightforward",str(patient.iloc[0])))
        except:
            print("Folder exists")
        patient_fc.to_csv(os.path.join(GENERAL,proj,"analysis/07_quantification/straightforward",str(patient.iloc[0]),filename),index=False)


for proj in projects:
    expressed = pd.read_csv(os.path.join(GENERAL,proj,"analysis/07_quantification/straightforward/tumor_1TPM.csv"))
    patients_dir = [ f.path for f in os.scandir(os.path.join(GENERAL,proj,"analysis/07_quantification/straightforward")) if f.is_dir() ]

    #### outfile for the common list

    outfile=os.path.join(GENERAL,proj,"analysis/07_quantification/straightforward/tumor_1TPM_n.csv")
    gene_df = pd.DataFrame(columns=['gene_name'])

    for p in patients_dir:
        p = str(p.split("/")[-1])
        try:
            csv = pd.read_csv(os.path.join(GENERAL,proj,"analysis/07_quantification/straightforward",p)+"/"+p+"_tumor1TPM.csv")
            gene_list = list()

            little_csv = csv[['gene_name']]
            gene_df = pd.concat([gene_df, little_csv], axis=0)
        except:
            continue

    #### count how many patients have each transcript
    gene_df = gene_df.merge(expressed, on=['gene_name'], how="inner")
    gene_df.drop_duplicates(inplace=True)
    gene_df['n'] = gene_df.groupby('gene_name')['gene_name'].transform('count')
    gene_df.drop_duplicates(inplace=True)
    gene_df.sort_values(by=['n'], ascending=False).to_csv(outfile, index=None)



Get 90% shared

all_dfs = pd.DataFrame(columns=['gene_id','gene_name','gene_type','Length','n','project'])

proj = "BLCA"
df = pd.read_csv(os.path.join(GENERAL,proj,"analysis/07_quantification/straightforward/tumor_1TPM_n.csv"))
df=df.replace('processed_pseudogene','lncRNA')
df=df.replace('protein_coding','Protein Coding')

top = df[df.n >= 2]
top.to_csv(os.path.join(GENERAL,proj,"analysis/07_quantification/straightforward/90shared_1TPM.csv"), index=None)
top['project'] = proj

all_dfs = pd.concat([top, all_dfs], axis=0)

Get GTFs 90% shared

for proj in projects:
    print(proj)
    ref = pd.read_csv("/users/genomics/sergiov/annotations_and_indexes/gencode.v41.primary_assembly.annotation.gtf", sep="\t", header=None, comment="#")
    ref['gene_id']=ref[8].str.split('gene_id "', expand=True)[1]
    ref['gene_id']=ref['gene_id'].str.split('"', expand=True)[0]
    ref['gene_id']=ref['gene_id'].str.split('.', expand=True)[0]

    full_90 = pd.DataFrame(columns = [0])
    #### novel
    CSV_REF=pd.read_csv(os.path.join(GENERAL,proj,"analysis/07_quantification/straightforward/90shared_1TPM.csv"))

    genes = CSV_REF.gene_id.values.tolist()
    full_90 = ref[ref['gene_id'].isin(genes)]
    full_90.drop('gene_id', axis=1, inplace=True)
    full_90.drop_duplicates(inplace=True)

    full_90.to_csv(os.path.join(GENERAL,proj,"analysis/07_quantification/straightforward/90shared.gtf"), sep="\t", header=None, index=None, quoting = 3)



Convert to fasta

%%bash -s "$GENERAL" "$bash_projects" "$GENOMEFASTA"

export PATH=/genomics/users/marta/tools/gffread-0.12.7.Linux_x86_64/:$PATH

for proj in $2; do
    echo $proj

    file=$1/$proj/analysis/07_quantification/straightforward/90shared.gtf

    #get fasta
    gffread --attrs gene_name,transcript_type,transcript_name -w ${file%%.*}.fa -g $3 $file

    #replace spaces by ;
    sed -i 's/\ /;/g' ${file%%.*}.fa
done

## Normal expressed

for proj in projects:
    fc = pd.read_csv(os.path.join(GENERAL,proj,"analysis/07_quantification/straightforward/TPMs_genenames.csv"))
    tumor_genes = list()

    patients=pd.read_csv(os.path.join(GENERAL,proj,"results/paired_patients.csv"),)

    for index,patient in patients.iterrows():
        patient_fc = fc[["gene_name",patient.iloc[1]]]

        tumor_patient_fc = patient_fc[patient_fc[patient.iloc[1]] >= 1 ]
        tumor_genes.extend(tumor_patient_fc.gene_name.values.tolist())

    tumor = fc[fc['gene_name'].isin(tumor_genes)]

    transcript_gene = pd.read_csv(os.path.join(GENOMEDIR,"transcript_gene_v41.txt"))
    transcript_gene.columns = ['gene_id', 'gene_name', 'gene_type']
    lncRNA_pseudo = ['lncRNA','processed_pseudogene']

    tumor_merged = tumor.merge(transcript_gene, on=['gene_name'], how="inner")

    lncRNA = tumor_merged[tumor_merged['gene_type'].isin(lncRNA_pseudo)]
    print("lncRNA: ",len(lncRNA))

    cds = tumor_merged[tumor_merged['gene_type'] == "protein_coding"]
    print("PROTEIN CODING: ",len(cds))


    tumor1FPKM = pd.concat([lncRNA, cds], ignore_index = True)
    tumor1FPKM.to_csv(os.path.join(GENERAL,proj,"analysis/07_quantification/straightforward/normal_1TPM_table_of_counts.csv"),index=False)
    tumor1FPKM[['gene_id','gene_name','gene_type']].to_csv(os.path.join(GENERAL,proj,"analysis/07_quantification/straightforward/normal_1TPM.csv"), index=False)

    tumor1FPKM
